requirements.txt
PySide6>=6.6.0,<7.0.0
Pillow>=10.2.0,<11.0.0
opencv-python>=4.9.0,<5.0.0
numpy>=1.26.0,<2.0.0
scipy>=1.11.0,<2.0.0
